refactor: mock validateId 핸들러 네이밍 변경 #894
